static castage

+++ source/MapPanel.cpp
@@ -1109,7 +1109,7 @@ void MapPanel::DrawTravelPlan()
 	{
 		next = player.TravelPlan()[i];
 
-		if(flagship->IsEnteringHyperspace() && (i == player.TravelPlan().size() - 1))
+		if(flagship->IsEnteringHyperspace() && (static_cast<int>(i == player.TravelPlan().size()) - 1))
 			for(auto &it : fuel)
 				if(it.first->IsEnteringHyperspace())
 					it.second += it.first->JumpNavigation().GetCheapestJumpType(previous, next).second;
@@ -1143,7 +1143,7 @@ void MapPanel::DrawTravelPlan()
 		Color drawColor = outOfFlagshipFuelRangeColor;
 		if(isWormhole)
 			drawColor = wormholeColor;
-		else if(i == (player.TravelPlan().size() - 1) && flagship->IsEnteringHyperspace())
+		else if(i == (static_cast<int>(player.TravelPlan().size()) - 1) && flagship->IsEnteringHyperspace())
 			drawColor = jumpInProgressColor;
 		else if(!stranded)
 			drawColor = withinFleetFuelRangeColor;
